requirements.txt
requests
python-binance
schedule
numpy
pandas
jupyterthemes
pymongo
dnspython
bson
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1